chore: nuke export.go
